Refactor testing setup.
